Added code to log disk errors to a new metric
Count of failed disks (if any) is now going to measurement called ScaleIOErrors. Uploaded a new JSON for the grafana template as well that includes it, but of course it's highly specific to my own cluster names here. If you've no disk errors of course, you won't see anything there.